reducing form int32 to int16 lasers consts

+++ g_laser.cpp
@@ -17,13 +17,13 @@ public:
 
 // Constants
 namespace LaserConstants {
-    constexpr int32_t MAX_LASERS = 6;
-    constexpr int32_t LASER_COST = 25;
-    constexpr int32_t LASER_INITIAL_DAMAGE = 1;
-    constexpr int32_t LASER_ADDON_DAMAGE = 4;
-    constexpr int32_t LASER_INITIAL_HEALTH = 125;
-    constexpr int32_t LASER_ADDON_HEALTH = 100;
-    constexpr int32_t MAX_LASER_HEALTH = 1750;
+    constexpr int16_t MAX_LASERS = 6;
+    constexpr int16_t LASER_COST = 25;
+    constexpr int16_t LASER_INITIAL_DAMAGE = 1;
+    constexpr int16_t LASER_ADDON_DAMAGE = 4;
+    constexpr int16_t LASER_INITIAL_HEALTH = 125;
+    constexpr int16_t LASER_ADDON_HEALTH = 100;
+    constexpr int16_t MAX_LASER_HEALTH = 1750;
     constexpr gtime_t LASER_SPAWN_DELAY = 1_sec;
     constexpr gtime_t LASER_TIMEOUT_DELAY = 150_sec;
     constexpr gtime_t TRACE_UPDATE_INTERVAL = 50_ms;
